pyproject.toml
[tool.poetry]
name = "room_booking"
version = "0.1.0"
description = "Room Booking Service"
authors = ["Ozy-Viking <admin@hankin.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
mysqlclient = "^2.2.0"
tqdm = "^4.66.1"
django-filter = "^23.3"
django-cors-headers = "^4.2.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
